refactor: move two screens to its package
